updated, more functionality, still needs a lot, but it can now create users
