+++ examples/keycloak-pepr/README.md
@@ -1,22 +1,13 @@
-# Pepr Module
+# Pepr Module for Keycloak initialization
 
 This is a Pepr Module. [Pepr](https://github.com/defenseunicorns/pepr) is a Kubernetes tranformation system
 written in Typescript.
 
-The `capabilities` directory contains all the capabilities for this module. By default,
-a capability is a single typescript file in the format of `capability-name.ts` that is
-imported in the root `pepr.ts` file as `import { HelloPepr } from "./capabilities/hello-pepr";`.
-Because this is typescript, you can organize this however you choose, e.g. creating a sub-folder
-per-capability or common logic in shared files or folders.
+## getting started
 
-Example Structure:
+You should have a kubernetes cluster ready with keycloak deployed (I started with the bitnami chart), this example currently requires running in VScode
 
-```
-Module Root
-├── package.json
-├── pepr.ts
-└── capabilities
-    ├── example-one.ts
-    ├── example-three.ts
-    └── example-two.ts
-```
+
+## module
+
+This module
