+++ examples/keycloak-pepr/capabilities/util.ts
@@ -1,0 +1,16 @@
+import fetch from 'node-fetch';
+
+export { generateRandomPassword  };
+  
+
+function generateRandomPassword(length: number): string {
+  const characters = 'ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789';
+  const charactersLength = characters.length;
+  let result = '';
+
+  for (let i = 0; i < length; i++) {
+    result += characters.charAt(Math.floor(Math.random() * charactersLength));
+  }
+
+  return result;
+}
